+++ appboy-component/src/ios-unified/Appboy_iOS_SDK.framework/Headers/ABKLocationManager.h
@@ -7,22 +7,16 @@ NS_ASSUME_NONNULL_BEGIN
 @interface ABKLocationManager : NSObject
 
 /*!
- * Use ABKDisableAutomaticLocationCollectionKey to disable automatic location tracking. 
+ * Use ABKEnableAutomaticLocationCollectionKey to enable automatic location tracking.
  * For more information, please refer to Appboy.h.
  */
-@property (readonly) BOOL disableLocationTracking;
+@property (readonly) BOOL enableLocationTracking;
 
 /*!
- * Calling this method gives Braze permission to request WhenInUse authorization on your behalf the next time we attempt to collect
- * location in the foreground.
+ * Use ABKEnableGeofencesKey to enable geofences.
+ * For more information, please refer to Appboy.h.
  */
-- (void)allowRequestWhenInUseLocationPermission;
-
-/*!
- * Calling this method gives Braze permission to request Always authorization on your behalf the next time we attempt to collect
- * significant location changes.
- */
-- (void)allowRequestAlwaysPermission;
+@property (readonly) BOOL enableGeofences;
 
 /*!
  * Calling this method will log a location using the regular location provider if a location is reported in under