+++ appboy-component/src/ios-unified/Appboy_iOS_SDK.framework/Headers/ABKNFBaseCardCell.h
@@ -1,8 +1,12 @@
 #import <UIKit/UIKit.h>
-#import <FLAnimatedImage/FLAnimatedImageView.h>
+#import <SDWebImage/SDAnimatedImageView+WebCache.h>
 #import "ABKCard.h"
 
-typedef void (^ABKOnCellHeightUpdateBlock)(void);
+@protocol ABKBaseNewsFeedCellDelegate <NSObject>
+
+- (void)refreshTableViewCellHeights;
+
+@end
 
 @interface ABKNFBaseCardCell : UITableViewCell
 
@@ -18,6 +22,8 @@ typedef void (^ABKOnCellHeightUpdateBlock)(void);
  */
 @property (weak, nonatomic) IBOutlet UIImageView *unreadIndicatorView;
 
+@property (weak, nonatomic) id <ABKBaseNewsFeedCellDelegate> delegate;
+
 /*!
  * Card root view related constraints
  */
@@ -25,11 +31,6 @@ typedef void (^ABKOnCellHeightUpdateBlock)(void);
 @property (weak, nonatomic) IBOutlet NSLayoutConstraint *rootViewTrailingConstraint;
 @property (weak, nonatomic) IBOutlet NSLayoutConstraint *rootViewTopConstraint;
 @property (weak, nonatomic) IBOutlet NSLayoutConstraint *rootViewBottomConstraint;
-
-/*!
- * This is a callback being called on height constraint change, so tableView can reload its UI state accordingly.
- */
-@property (nonatomic, copy) ABKOnCellHeightUpdateBlock onCellHeightUpdateBlock;
 
 /*!
  * These are basic UI configuration for the News Feed. They are set to the default value in `setUp`